Complete splitting of visitor.
